requirements.txt
fastapi
uvicorn
llama-index
pydantic
python-dotenv
beautifulsoup4
pandas
PyPDF2
prometheus-client
requests
langchain
google-generativeai
schedule